requirements.txt
streamlit
pandas
requests
stripe
supabase
postgrest
realtime
storage3
fastapi
uvicorn
python-multipart
python-dotenv